frontend/requirements.txt
streamlit==1.28.1
streamlit-chat==0.1.1
python-socketio[client]==5.10.0
requests==2.31.0
extra-streamlit-components==0.1.60